Added a new resource
Hey, I have added a reference link for an array of objects under Description. I came upon this article while looking for resources to learn Sort an Array of Objects in JavaScript. This citation, in my opinion, will enhance the content of this article. Hope that my contribution will benefit other learners.